+++ apps/web/client/src/app/project/[id]/page.tsx
@@ -11,10 +11,10 @@ async function getProject(id: string) {
 
     const newFrame: WebFrame = {
         id: '1',
-        //url: 'https://3cczd6-8084.csb.app',
-        url: 'http://localhost:8084',
+        url: 'https://3cczd6-8084.csb.app',
+        // url: 'http://localhost:8084',
         position: { x: 200, y: 200 },
-        dimension: { width: 1440, height: 900 },
+        dimension: { width: 500, height: 500 },
         type: FrameType.WEB,
     };
