--- apps/web/client/src/components/store/editor/engine/code/diff.ts
@@ -1,20 +1,0 @@
-import type { CodeDiff, FileToRequests } from '@onlook/models';
-import { getAstFromContent, getContentFromAst, transformAst } from "@onlook/parser";
-
-export async function processGroupedRequests(groupedRequests: FileToRequests): Promise<CodeDiff[]> {
-    const diffs: CodeDiff[] = [];
-    for (const [path, request] of groupedRequests) {
-        const { oidToRequest, content } = request;
-        const ast = getAstFromContent(content);
-
-        if (!ast) {
-            continue;
-        }
-
-        const original = await getContentFromAst(ast);
-        transformAst(ast, oidToRequest);
-        const generated = await getContentFromAst(ast);
-        diffs.push({ original, generated, path });
-    }
-    return diffs;
-}
